+++ packages/tools/buildServerDocker.ts
@@ -1,5 +1,6 @@
-import { execCommand2, rootDir } from './tool-utils';
+import { rootDir } from './tool-utils';
 import * as moment from 'moment';
+import { execCommand } from '@joplin/utils';
 
 interface Argv {
 	dryRun?: boolean;
@@ -35,7 +36,7 @@ async function main() {
 	const buildDate = moment(new Date().getTime()).format('YYYY-MM-DDTHH:mm:ssZ');
 	let revision = '';
 	try {
-		revision = await execCommand2('git rev-parse --short HEAD', { showStdout: false });
+		revision = await execCommand('git rev-parse --short HEAD', { showStdout: false });
 	} catch (error) {
 		console.info('Could not get git commit: metadata revision field will be empty');
 	}
@@ -62,11 +63,11 @@ async function main() {
 		return;
 	}
 
-	await execCommand2(dockerCommand);
+	await execCommand(dockerCommand);
 
 	for (const tag of dockerTags) {
-		await execCommand2(`docker tag "${repository}:${imageVersion}" "${repository}:${tag}"`);
-		if (pushImages) await execCommand2(`docker push ${repository}:${tag}`);
+		await execCommand(`docker tag "${repository}:${imageVersion}" "${repository}:${tag}"`);
+		if (pushImages) await execCommand(`docker push ${repository}:${tag}`);
 	}
 }