+++ packages/tools/release-android.ts
@@ -1,5 +1,6 @@
+import { execCommand } from '@joplin/utils';
 import * as fs from 'fs-extra';
-import { execCommandVerbose, execCommandWithPipes, githubRelease, githubOauthToken, fileExists, gitPullTry, completeReleaseWithChangelog, execCommand2 } from './tool-utils';
+import { execCommandVerbose, execCommandWithPipes, githubRelease, githubOauthToken, fileExists, gitPullTry, completeReleaseWithChangelog } from './tool-utils';
 const path = require('path');
 const fetch = require('node-fetch');
 const uriTemplate = require('uri-template');
@@ -150,7 +151,7 @@ async function main() {
 	const isPreRelease = !('type' in argv) || argv.type === 'prerelease';
 
 	process.chdir(rnDir);
-	await execCommand2('yarn run build', { showStdout: false });
+	await execCommand('yarn run build', { showStdout: false });
 
 	if (isPreRelease) console.info('Creating pre-release');
 	console.info('Updating version numbers in build.gradle...');